task_manager/requirements.txt
psycopg2
flask
google-auth 
google-auth-oauthlib
google-auth-httplib2 
google-api-python-client
google-cloud-pubsub
